src/qiki/services/operator_console/requirements.txt
# Core dependencies
rich>=13.7.0
textual>=0.47.1
nats-py>=2.6.0
grpcio>=1.60.0
protobuf>=4.25.1
pydantic>=2.5.0
python-dotenv>=1.0.0
psutil>=5.9.0

# Testing dependencies
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-cov>=4.0.0
pytest-mock>=3.10.0
pytest-xdist>=3.0.0
coverage>=7.0.0

# Development and linting
ruff>=0.1.0
pylint>=2.15.0
mypy>=1.0.0
black>=22.0.0
isort>=5.10.0
pre-commit>=2.20.0

# Dev dependencies (для тестов и качества кода внутри контейнера)
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
ruff>=0.1.9
pylint>=3.0.3
pre-commit>=3.5.0